git-svn-id: https://svn.neo4j.org/components/neo/trunk@2846 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ community/src/java/org/neo4j/impl/nioneo/xa/NeoStoreXaDataSource.java
@@ -332,7 +332,7 @@ public class NeoStoreXaDataSource extends XaDataSource
         return store.nextId();
     }
 
-    public int getHighestPossibleIdInUse( Class<?> clazz )
+    public long getHighestPossibleIdInUse( Class<?> clazz )
     {
         Store store = idGenerators.get( clazz );
         if ( store == null )
@@ -343,7 +343,7 @@ public class NeoStoreXaDataSource extends XaDataSource
         return store.getHighestPossibleIdInUse();
     }
 
-    public int getNumberOfIdsInUse( Class<?> clazz )
+    public long getNumberOfIdsInUse( Class<?> clazz )
     {
         Store store = idGenerators.get( clazz );
         if ( store == null )
